% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map of accidents from FARS data}
\source{
FARS data can be from 
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer or character representing the number 
corresponding to the State in the US for analysis.}

\item{year}{A character string or numeric representing the year of the FARS 
data file required for further processing.}
}
\value{
A map plot of accident locations, represented as points, for a given 
US State and year with information from the appropriate FARS data file. The 
State boundary will also be drawn.
}
\description{
Given a US State number and a year, this function will produce a simple black 
and white map indicating accident locations as contained in a FARS data file.
}
\note{
Conditions that may result in an error include:
\itemize{
  \item This function uses \code{\link{make_filename}} and 
    \code{\link{fars_read}} and these must be loaded into the NAMESPACE.
  \item This function also requires the packages \code{\link{dplyr}} and 
    \code{\link{maps}} to be installed and loaded into the NAMESPACE.
  \item Providing an incorrect US State number.
  \item The year input must be coerceable to class integer.
  \item The provided year must correspond to a downloaded FARS data file.
  \item The function must be run from the same location as the downloaded 
    FARS files that require analysis.
  }
}
\examples{
\dontrun{
fars_map_state(47, 2013)
fars_map_state("47", "2013")
}

}
