% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data into a data frame.}
\source{
FARS data can be from 
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the name of a download FARS data file. 
String can include path to the file if the file to be read is not in the 
working directory.}
}
\value{
This function unzips the FARS data and returns a data frame of class
   'tbl_df', 'tbl' and 'data.frame'.
}
\description{
\code{fars_read} is a helper function designed to read data downloaded from 
the Fatality Analysis Reporting System (FARS).
}
\note{
Conditions that may result in an error include:
\itemize{
  \item \code{fars_read} uses functions from the packages 
    \code{\link[dplyr]{tbl_df}},and  \code{\link[readr]{read_csv}} and they 
    therefore must be installed prior to use. 
  \item \code{fars_read} also expects the character string of the FARS data 
    file name to be the full file name (i.e. include the file extension).
    }
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")}

}
