% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of FARS data into a list of data frames.}
\source{
FARS data can be from 
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character or numeric vector representing the years of the FARS 
data files required for further processing.}
}
\value{
The function will return a list containing as many data frames as 
years provided as input. The list is of class 'list' whilst the individual 
data frames are of class 'tbl_df', 'tbl' and 'data.frame'. Each data frame 
contains the columns:
\describe{
 \item{MONTH}{The months corresponding to each observation in the FARS data 
 file}
 \item{year}{The year corresponding to each observation in the FARS data file}}
}
\description{
Given a vector of years corresponding to Fatality Analysis Reporting System 
(FARS) data file names,  \code{fars_read_years} will read each of the FARS 
data files and create a list of data frames of the months and years contained 
in each FARS data file.
}
\note{
Conditions that may result in an error include:
\itemize{
  \item This function uses \code{\link{make_filename}} and as such it must be 
    loaded into the NAMESPACE.
  \item This function and others that it rely on, \code{\link{make_filename}},
    require the packages \code{\link{dplyr}} and \code{\link{readr}} to be 
    installed.
  \item This function also requires that the package \code{\link{dpyr}} is 
    loaded into the NAMESPACE.
  \item This function must be run from the same location as the downloaded
    FARS data files.
  \item Providing a vector of years where there is no corresponding FARS data
    file will result in an error for that year/years. The function will still 
    return a list however, and it will contain results for years that do have 
    a corresponding FARS data file.
  }
}
\examples{
 \dontrun{
 fars_read_years(c(2013, 2014, 2015))
 fars_read_years(c("2013", "2014", "2015"))
 }
 
}
