% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise FARS data.}
\source{
FARS data can be from 
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character or numeric vector representing the years of the FARS 
data files required for further processing.}
}
\value{
This function will return a data frame of class 'tbl_df', 'tbl' and 
'data.frame'. The data frame will contain columns:
\describe{
 \item{MONTH}{Numerical month abbreviations 1-12}
 \item{YYYY}{'YYYY' represents a named year. There will be a separate column 
 for each year as provided in the vector of years as input. Each column will 
 contain the number of fatalities per month for that year.}
 }
}
\description{
Given a vector of years corresponding to Fatality Analysis Reporting System 
(FARS) data file names, \code{fars_summarize_years} will read each of the FARS 
data files and create a single data frame summarising fatalities by month for 
the input years.
}
\note{
Conditions that may result in an error include:
\itemize{
  \item This function uses \code{\link{make_filename}} and 
    \code{link{fars_read_years}}. As such they must be loaded into the 
    NAMESPACE.
  \item This function and others that it rely on, \code{\link{make_filename}}
    and \code{link{fars_read_years}}, require the packages \code{\link[dplyr]{summarise}}, 
    \code{\link[readr]{read_csv}} and \code{\link[tidyr]{spread}} to be installed.
  \item This function must be run from the same location as the downloaded
    FARS data files.
  \item Providing a vector of years where there is no corresponding FARS data
    file will result in an error for that year/years. The function will still 
    return a data frame however, and it will contain results for years that 
    do have a corresponding FARS data file.
  }
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years(c("2013", "2014", "2015"))
}

}
