% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make generic FARS data file name.}
\source{
FARS data can be from 
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{here}.
}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or numeric representing the year of the FARS 
data file required for further processing.}
}
\value{
This function returns a correctly formatted FARS data file name as a
character string given the year of interest.
}
\description{
\code{make_filename} is a helper function that constructs the correct 
Fatality Analysis Reporting System (FARS) data file name given a year. This 
function can be used to generate the input required for the 
\code{\link{fars_read}} function.
}
\note{
Conditions that may result in an error include:
\itemize{
  \item \code{make_filename} will not recognise inputs that cannot be coerced
  to an integer. For example 2013 and "2013" are acceptable inputs whereas
  "two thousand and thirteen" will cause an error.
  }
}
\examples{
make_filename(2013)
make_filename("2013")

}
